requirements.txt
altgraph==0.17.4
blinker==1.8.2
bottle==0.13.1
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
macholib==1.16.3
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
proxy_tools==0.1.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-WebKit==10.3.1
pywebview==5.2
setuptools==75.1.0
typing_extensions==4.12.2
Werkzeug==3.0.4
altgraph==0.17.4
blinker==1.8.2
bottle==0.13.1
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
macholib==1.16.3
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
proxy_tools==0.1.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-WebKit==10.3.1
pywebview==5.2
setuptools==75.1.0
typing_extensions==4.12.2
Werkzeug==3.0.4
altgraph==0.17.4
blinker==1.8.2
bottle==0.13.1
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
macholib==1.16.3
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
proxy_tools==0.1.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-WebKit==10.3.1
pywebview==5.2
setuptools==75.1.0
typing_extensions==4.12.2
Werkzeug==3.0.4
